libunit-wasm: Fix rust build
The rust build failed when trying to publish new crates, although for
some reason it didn't fail before that, due to the luw_srb_flags_t enums
argument to luw_set_req_buf() being an unsigned long (which is 64 bits
on my system) and thus producing a type mismatch error.

Rather than fight with rust just make the flags argument an unsigned
int, 32 bits is more than enough for this anyway.

+++ src/c/include/unit/unit-wasm.h
@@ -161,7 +161,7 @@ extern u32 luw_malloc_handler(size_t size);
 #pragma GCC visibility push(default)
 
 extern void luw_init_ctx(luw_ctx_t *ctx, u8 *addr, size_t offset);
-extern int luw_set_req_buf(luw_ctx_t *ctx, u8 **buf, unsigned long flags);
+extern int luw_set_req_buf(luw_ctx_t *ctx, u8 **buf, unsigned int flags);
 extern const char *luw_get_http_path(const luw_ctx_t *ctx);
 extern const char *luw_get_http_method(const luw_ctx_t *ctx);
 extern const char *luw_get_http_version(const luw_ctx_t *ctx);
